+++ src/Plugin.php
@@ -1,24 +1,38 @@
 <?php
-namespace enovate\rollbar;
+namespace enovatedesign\rollbar;
+
+use enovatedesign\rollbar\services\Rollbar as RollbarService;
+use enovatedesign\rollbar\variables\Rollbar as RollbarVariable;
+
+use Rollbar\Payload\Level;
 
 use Craft;
-use craft\events\DefineComponentsEvent;
+use craft\base\Plugin as BasePlugin;
 use craft\events\ExceptionEvent;
-use craft\events\RegisterUrlRulesEvent;
 use craft\web\ErrorHandler;
 use craft\web\twig\variables\CraftVariable;
-use craft\web\UrlManager;
-use enovate\rollbar\services\Rollbar as RollbarService;
-use enovate\rollbar\variables\Rollbar as RollbarVariable;
-use Rollbar\Payload\Level;
 use yii\base\Event;
 
-class Plugin extends \craft\base\Plugin
+/**
+ * Class Rollbar
+ * @package enovate\rollbar
+ * @since 0.1.0
+ *
+ * @property RollbarService $rollbar
+ */
+class Plugin extends BasePlugin
 {
-    public $hasCpSettings = true;
+    /**
+     * @var Plugin
+     */
+    public static $plugin;
 
     public function init()
     {
+        parent::init();
+
+        self::$plugin = $this;
+
         $this->setComponents([
             'rollbar' => RollbarService::class,
         ]);
@@ -31,17 +45,12 @@ class Plugin extends \craft\base\Plugin
             }
         });
 
-        Event::on(CraftVariable::class, CraftVariable::EVENT_DEFINE_COMPONENTS, function(DefineComponentsEvent $event)
+        Event::on(CraftVariable::class, CraftVariable::EVENT_INIT, function(Event $event)
         {
-            $event->components['rollbar'] = new RollbarVariable();
+            /** @var RollbarVariable $variable */
+            $variable = $event->sender;
+            $variable->set('rollbar', RollbarVariable::class);
         });
-
-        Event::on(UrlManager::class, UrlManager::EVENT_REGISTER_CP_URL_RULES, function(RegisterUrlRulesEvent $event)
-        {
-            $event->rules['rollbar/csp'] = 'rollbar/csp/view';
-        });
-
-        parent::init();
     }
 
     /**
@@ -55,12 +64,12 @@ class Plugin extends \craft\base\Plugin
     public function settingsHtml()
     {
         $configFile = Craft::$app->getConfig()->getConfigFromFile('rollbar');
-        $settings   = $this->getSettings();
+        $settings = $this->getSettings();
 
         $settings->validate();
 
         return Craft::$app->getView()->renderTemplate('rollbar/_settings', [
-            'settings'   => $settings,
+            'settings' => $settings,
             'configFile' => $configFile,
         ]);
     }