+++ src/config.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'accessToken' => 'abc123',
+    'clientAccessToken' => 'abc123',
+    'environment' => 'devlopment',
+    'jsTracking' => false,
+    'reporting' => true,
+];