+++ src/models/Settings.php
@@ -1,5 +1,5 @@
 <?php
-namespace enovate\rollbar\models;
+namespace enovatedesign\rollbar\models;
 
 use craft\base\Model;
 use craft\helpers\UrlHelper;
@@ -11,7 +11,6 @@ class Settings extends Model
     public $reporting;
     public $jsTracking;
     public $captureUnhandledRejections;
-    public $cspEndpoint;
     public $ignoreHTTPCodes;
     public $environment;
 
@@ -20,7 +19,6 @@ class Settings extends Model
         $this->reporting                  = true;
         $this->jsTracking                 = true;
         $this->captureUnhandledRejections = false;
-        $this->cspEndpoint                = false;
         $this->ignoreHTTPCodes            = '404, 403, 503';
         $this->environment                = UrlHelper::siteUrl();
 
@@ -35,7 +33,7 @@ class Settings extends Model
         return [
             [['accessToken', 'environment'], 'required'],
             [['accessToken', 'clientAccessToken'], 'string', 'length' => 32],
-            [['reporting', 'jsTracking', 'captureUnhandledRejections', 'cspEndpoint'], 'boolean'],
+            [['reporting', 'jsTracking', 'captureUnhandledRejections'], 'boolean'],
         ];
     }
 }