+++ src/translations/en/rollbar.php
@@ -1,0 +1,17 @@
+<?php
+
+return [
+    'Access Token' => '',
+    'Test' => '',
+    'Client Access Token' => '',
+    'Ignore Error Codes' => '',
+    'This is a test warning' => '',
+    'This is a test error' => '',
+    'Environment' => '',
+    'Track JavaScript Errors' => '',
+    'Test message posted successfully' => '',
+    'Test message was not posted successfully' => '',
+    'Your settings are invalid' => '',
+    'This is being set by the {setting} config setting in <code>config/rollbar.php</code>.' => '',
+    'Reporting is currently disabled for this environment (<code>{environment}</code>).' => '',
+];